done pub sub

+++ apps/webhooks/views.py
@@ -1,3 +1,4 @@
+from typing import Optional
 from django.contrib.contenttypes.models import ContentType
 from rest_framework import status
 from rest_framework.generics import get_object_or_404
@@ -16,11 +17,13 @@ class HookObserverModelViewSet(ModelViewSet):
     queryset = HookObserver.objects.all()
     serializer_class = HookObserverModelSerializer
 
-    def find_hook_observer_model(self, request_data):
+    def find_hook_observer_model(self, request_data) -> Optional[dict]:
         ser = HookObserverSubscribeSerializer(data=request_data)
         ser.is_valid(raise_exception=True)
         data = ser.data
         data_ct = data.get('content_type')
+        if not data_ct:
+            return None
         ct = get_object_or_404(ContentType,
             app_label=data_ct.get('app_label'),
             model=data_ct.get('model')
@@ -28,22 +31,30 @@ class HookObserverModelViewSet(ModelViewSet):
         find_kwargs = {
             'notification_type': data.get('notification_type'),
             'client_url': data.get('client_url'),
-             'content_type': ct
+            'content_type': ct
         }
         return find_kwargs
 
     @action(methods=['put'], detail=False)
     def subscribe(self, request):
         find_kwargs = self.find_hook_observer_model(request.data)
+        if not find_kwargs:
+            return Response('Hook observer model not found', status=status.HTTP_404_NOT_FOUND)
         ho, created = HookObserver.objects.get_or_create(**find_kwargs)
 
         serializer = self.serializer_class(instance=ho)
         headers = self.get_success_headers(serializer.data)
-        return Response(serializer.data, status=status.HTTP_201_CREATED, headers=headers)
+        return Response(
+            serializer.data,
+            status=status.HTTP_201_CREATED if created else status.HTTP_200_OK,
+            headers=headers
+        )
 
     @action(methods=['put'], detail=False)
     def unsubscribe(self, request):
         find_kwargs = self.find_hook_observer_model(request.data)
+        if not find_kwargs:
+            return Response('Hook observer model not found', status=status.HTTP_404_NOT_FOUND)
         ho = get_object_or_404(HookObserver, **find_kwargs)
         ho.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
